requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
backcall==0.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.47.2
fsspec==2023.12.2
gensim==3.8.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
ijson==3.2.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
JPype1==1.5.0
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
konlpy==0.6.0
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mpmath==1.3.0
namu-wiki-extractor==0.3.0
nest-asyncio==1.5.9
networkx==3.1
nltk==3.8.1
numpy==1.24.4
openai==1.7.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
stack-data==0.6.3
sympy==1.12
tokenizers==0.15.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0
